flex_entry structures unified and simplified. dirname in flex_named_key made into a 0-terminated string.
Attention: Due to a bug in the WebIDE (https://github.com/kaitai-io/kaitai_struct/issues/246), clicking on "block_id" or "record" leads to an error msg if that's the wrong type for the node's alignment type. It's not harmful, though - one can keep using the IDE after the error is displayed.