Turning _type suffixes into prefixes, improving EA handling (#14)
* Update apfs.ksy

* Adds special case for thread_record variant where another 4 bytes are inserted between name_length and name.

* Adds item_type of value 10

* volume and file names are now properly handled as NUL-terminated strings.

* Removed unused "flex_key" structure.

* Renaming a few ids to meet the ksy style guide (http://doc.kaitai.io/ksy_style_guide.html): len_name instead of name_len etc.

* flex_entry structures unified and simplified.
dirname in flex_named_key made into a 0-terminated string.

Attention: Due to a bug in the WebIDE (https://github.com/kaitai-io/kaitai_struct/issues/246), clicking on "block_id" or "record" leads to an error msg if that's the wrong type for the node's alignment type. It's not harmful, though - one can keep using the IDE after the error is displayed.

* Removed unused "flex_key" structure.

* Renaming a few ids to meet the ksy style guide (http://doc.kaitai.io/ksy_style_guide.html): len_name instead of name_len etc.

* flex_entry structures unified and simplified.
dirname in flex_named_key made into a 0-terminated string.

Attention: Due to a bug in the WebIDE (https://github.com/kaitai-io/kaitai_struct/issues/246), clicking on "block_id" or "record" leads to an error msg if that's the wrong type for the node's alignment type. It's not harmful, though - one can keep using the IDE after the error is displayed.

* Rename entry_locations back to entries

* Adds flex_extattr_record

* Adds proper case for symlink EA records.
Renamed ids xxx_type into type_xxx.
Renamed node_type to content_type/type_content.
Renamed alignment_type to node_type/type_node because it specifies the format of all the entries.

* No more warnings in generated C++ code by adding size or default cases to all switch clauses.

* Fixes a syntax error for stricter yaml parsers